Update necessary imports
