+++ src/plugin/plugin.ts
@@ -1,14 +1,14 @@
 import type { Node, Schema } from "prosemirror-model";
 import { Plugin } from "prosemirror-state";
 import type { EditorProps } from "prosemirror-view";
-import type { FieldNameToValueMap } from "./fieldViews/helpers";
-import type { Commands } from "./helpers";
-import { createDecorations } from "./helpers";
 import type {
   ElementSpec,
   FieldNameToFieldViewSpec,
   FieldSpec,
-} from "./plugin/types/Element";
+} from "../plugin/types/Element";
+import type { FieldNameToValueMap } from "./fieldViews/helpers";
+import type { Commands } from "./helpers";
+import { createDecorations } from "./helpers";
 import { getElementFieldViewFromType } from "./pluginHelpers";
 
 const decorations = createDecorations("imageElement");
